add to main result file
